Add neovim command to restart Pyright
Pyright doesn't detect when a file is added to the workspace because
"The workspace/didChangeWatchedFiles LSP client capability is now
enabled by default on Mac and Windows. Disabled on Linux since there
currently isn't a viable backend for watching files that scales well for
large directories.", as seen here:
https://neovim.io/doc/user/news-0.10.html

When I add a new python file and import it, the LSP gives a diagonstic
error cause it cannot find the file. Running PyrightRestart will fix
this.